Fix gradle build
