+++ idl-declarations/src/main/antlr4/org/antlr/webidl/WebIDL.g4
