+++ idl-declarations/src/org/jetbrains/dukat/idlDeclarations/IDLDeclaration.kt
@@ -1,0 +1,5 @@
+package org.jetbrains.dukat.idlDeclarations
+
+import org.jetbrains.dukat.astCommon.TopLevelEntity
+
+interface IDLDeclaration : TopLevelEntity
