+++ idl-declarations/src/org/jetbrains/dukat/idlDeclarations/InterfaceIDLDeclaration.kt
@@ -1,0 +1,3 @@
+package org.jetbrains.dukat.idlDeclarations
+
+class InterfaceIDLDeclaration(val name: String) : IDLDeclaration
